fix(backend): bigint Long 타입 회귀 복원 + spotbugs 2건 수정
prettier-lint-rename 머지(#210)가 bigint 이전 코드 기반이라, b387259(#211)의
Long 타입을 ~87개 파일에서 Integer로 되돌려 컴파일이 깨졌다
(CI-backend bootJar / schema-validate SpringBoot 동시 실패의 단일 근본 원인).
b387259의 Long 로직을 복원하고 googleJavaFormat(spotless)으로 재포맷.

spotbugs 2건도 해결(ignoreFailures=false 유지):
- LegacyPbkdf2Verifier: catch(Exception) → catch(GeneralSecurityException | RuntimeException)
- ScriptService: 삼항 Double 언박싱→재박싱 제거 (Double.valueOf)

+++ backend/springboot/src/main/java/com/workmaite/domain/agendas/controller/AgendaController.java
@@ -26,7 +26,7 @@ public class AgendaController {
   // AI가 회의체 자료를 분석해 안건을 자동 추출
   @PostMapping("/meetings/{meetingId}/agendas/extract")
   public ResponseEntity<ApiResponse<List<AgendaResponse>>> extractAgendas(
-      @PathVariable Integer meetingId, @RequestBody @Valid AgendaExtractRequest request) {
+      @PathVariable Long meetingId, @RequestBody @Valid AgendaExtractRequest request) {
     return ResponseEntity.status(HttpStatus.CREATED)
         .body(ApiResponse.ok(agendaService.extractAgendas(meetingId, request)));
   }
@@ -34,34 +34,34 @@ public class AgendaController {
   // 회의체의 안건 목록을 orderIndex 오름차순으로 반환
   @GetMapping("/meetings/{meetingId}/agendas")
   public ResponseEntity<ApiResponse<List<AgendaResponse>>> getAgendas(
-      @PathVariable Integer meetingId) {
+      @PathVariable Long meetingId) {
     return ResponseEntity.ok(ApiResponse.ok(agendaService.getAgendas(meetingId)));
   }
 
   // 안건 수동 생성
   @PostMapping("/meetings/{meetingId}/agendas")
   public ResponseEntity<ApiResponse<AgendaResponse>> createAgenda(
-      @PathVariable Integer meetingId, @RequestBody @Valid AgendaCreateRequest request) {
+      @PathVariable Long meetingId, @RequestBody @Valid AgendaCreateRequest request) {
     return ResponseEntity.status(HttpStatus.CREATED)
         .body(ApiResponse.ok(agendaService.createAgenda(meetingId, request)));
   }
 
   // 안건 단건 조회
   @GetMapping("/agendas/{agendaId}")
-  public ResponseEntity<ApiResponse<AgendaResponse>> getAgenda(@PathVariable Integer agendaId) {
+  public ResponseEntity<ApiResponse<AgendaResponse>> getAgenda(@PathVariable Long agendaId) {
     return ResponseEntity.ok(ApiResponse.ok(agendaService.getAgenda(agendaId)));
   }
 
   // 안건 수정 (title, content, orderIndex, status 부분 수정)
   @PatchMapping("/agendas/{agendaId}")
   public ResponseEntity<ApiResponse<AgendaResponse>> updateAgenda(
-      @PathVariable Integer agendaId, @RequestBody AgendaUpdateRequest request) {
+      @PathVariable Long agendaId, @RequestBody AgendaUpdateRequest request) {
     return ResponseEntity.ok(ApiResponse.ok(agendaService.updateAgenda(agendaId, request)));
   }
 
   // 안건 삭제
   @DeleteMapping("/agendas/{agendaId}")
-  public ResponseEntity<ApiResponse<Void>> deleteAgenda(@PathVariable Integer agendaId) {
+  public ResponseEntity<ApiResponse<Void>> deleteAgenda(@PathVariable Long agendaId) {
     agendaService.deleteAgenda(agendaId);
     return ResponseEntity.ok(ApiResponse.ok(null));
   }
@@ -69,7 +69,7 @@ public class AgendaController {
   // 안건 담당자 배정
   @PatchMapping("/agendas/{agendaId}/assignment")
   public ResponseEntity<ApiResponse<AgendaResponse>> assignAgenda(
-      @PathVariable Integer agendaId, @RequestBody @Valid AgendaAssignmentRequest request) {
+      @PathVariable Long agendaId, @RequestBody @Valid AgendaAssignmentRequest request) {
     return ResponseEntity.ok(ApiResponse.ok(agendaService.assignAgenda(agendaId, request)));
   }
 }
