+++ backend/springboot/src/main/java/com/workmaite/domain/chat/controller/ChatMessageController.java
@@ -22,9 +22,9 @@ public class ChatMessageController {
   public ResponseEntity<ApiResponse<List<ChatMessageResponse>>> getChatHistory(
       Authentication authentication,
       @RequestParam String threadId,
-      @RequestParam(required = false) Integer beforeId,
+      @RequestParam(required = false) Long beforeId,
       @RequestParam(required = false) Integer limit) {
-    Integer userId = Integer.parseInt(authentication.getName());
+    Long userId = Long.parseLong(authentication.getName());
     if (!isAccessible(threadId, userId)) {
       return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ApiResponse.fail("접근 권한이 없습니다."));
     }
@@ -36,7 +36,7 @@ public class ChatMessageController {
   @DeleteMapping("/chat/messages")
   public ResponseEntity<ApiResponse<Void>> clearChatHistory(
       Authentication authentication, @RequestParam String threadId) {
-    Integer userId = Integer.parseInt(authentication.getName());
+    Long userId = Long.parseLong(authentication.getName());
     if (!isAccessible(threadId, userId)) {
       return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ApiResponse.fail("접근 권한이 없습니다."));
     }
@@ -48,11 +48,11 @@ public class ChatMessageController {
    * global_{userId}, archive_{userId} 스레드는 본인만 접근 가능. meeting_{meetingId}, session_{sessionId} 스레드는
    * 인증된 사용자에게 허용 (AI 백엔드가 멤버 권한 체크).
    */
-  private boolean isAccessible(String threadId, Integer userId) {
+  private boolean isAccessible(String threadId, Long userId) {
     if (threadId.startsWith("global_") || threadId.startsWith("archive_")) {
       String prefix = threadId.startsWith("global_") ? "global_" : "archive_";
       try {
-        return Integer.parseInt(threadId.substring(prefix.length())) == userId;
+        return Long.parseLong(threadId.substring(prefix.length())) == userId;
       } catch (NumberFormatException e) {
         return false;
       }
