+++ backend/springboot/src/main/java/com/workmaite/domain/meetings/repository/MeetingRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MeetingRepository extends JpaRepository<Meeting, Integer> {
+public interface MeetingRepository extends JpaRepository<Meeting, Long> {
 
   List<Meeting> findByTitleContainingIgnoreCaseOrDescriptionContainingIgnoreCase(
       String title, String description);
@@ -16,16 +16,16 @@ public interface MeetingRepository extends JpaRepository<Meeting, Integer> {
   @Query(
       "SELECT m FROM Meeting m WHERE m.id IN (SELECT mm.meetingId FROM MeetingMember mm WHERE mm.userId = :userId) AND m.status = :status")
   List<Meeting> findByUserIdAndStatus(
-      @Param("userId") Integer userId, @Param("status") MeetingStatus status);
+      @Param("userId") Long userId, @Param("status") MeetingStatus status);
 
   @Query(
       "SELECT m FROM Meeting m WHERE m.id IN (SELECT mm.meetingId FROM MeetingMember mm WHERE mm.userId = :userId)")
-  List<Meeting> findByUserId(@Param("userId") Integer userId);
+  List<Meeting> findByUserId(@Param("userId") Long userId);
 
   // P8-4: 페이지네이션 변형 (파라미터 없는 기존 메서드는 호환용으로 유지)
   @Query(
       "SELECT m FROM Meeting m WHERE m.id IN (SELECT mm.meetingId FROM MeetingMember mm WHERE mm.userId = :userId)")
-  List<Meeting> findByUserId(@Param("userId") Integer userId, Pageable pageable);
+  List<Meeting> findByUserId(@Param("userId") Long userId, Pageable pageable);
 
   List<Meeting> findByTitleContainingIgnoreCaseOrDescriptionContainingIgnoreCase(
       String title, String description, Pageable pageable);
