+++ backend/springboot/src/main/java/com/workmaite/domain/reports/dto/ReportSessionSubmitRequest.java
@@ -10,10 +10,10 @@ import lombok.NoArgsConstructor;
 public class ReportSessionSubmitRequest {
 
   @NotNull(message = "회의체 ID를 입력해주세요.")
-  private Integer meetingId;
+  private Long meetingId;
 
   @NotNull(message = "업로더 ID를 입력해주세요.")
-  private Integer uploadId;
+  private Long uploadId;
 
   @NotBlank(message = "제출 부서를 입력해주세요.")
   private String submitterDepartment;
