+++ backend/springboot/src/main/java/com/workmaite/domain/reports/repository/ReportRepository.java
@@ -4,7 +4,7 @@ import com.workmaite.domain.reports.entity.Report;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Report, Integer> {
+public interface ReportRepository extends JpaRepository<Report, Long> {
 
-  List<Report> findAllByMeetingId(Integer meetingId);
+  List<Report> findAllByMeetingId(Long meetingId);
 }
